dokończone zaJ5

+++ zaj2/src/main/java/pl/edu/pjatk/zaj2/service/MyRestService.java
@@ -42,6 +42,9 @@ public class MyRestService {
 
     public void remove(String name) {
         List<Zwierze> zwierze = this.repository.findByName(name);
+        if (zwierze.isEmpty()) {
+            throw new ResorceNotExistException();
+        }
         this.repository.deleteAll(zwierze);
     }
